+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 Released on 30/03/2025
 
 - [issue 16](https://github.com/veeso/tuifeed/issues/16): filter out old articles from the history.
+- [issue 17](https://github.com/veeso/tuifeed/issues/17): mark as all read feature. Use `V` to mark all articles from a **source** as read. Use `CTRL+V` to mark all articles from **all sources** as read.
 - [issue 18](https://github.com/veeso/tuifeed/issues/18): first article is not marked as read when first loaded. Also mark source as read if all articles are read
 
 ## 0.4.1